+++ lib/src/code_generator.dart
@@ -11,6 +11,7 @@ export 'code_generator/constant.dart';
 export 'code_generator/enum_class.dart';
 export 'code_generator/func.dart';
 export 'code_generator/global.dart';
+export 'code_generator/imports.dart';
 export 'code_generator/library.dart';
 export 'code_generator/struc.dart';
 export 'code_generator/type.dart';