+++ lib/src/code_generator/func.dart
@@ -4,9 +4,7 @@
 
 import 'package:ffigen/src/code_generator.dart';
 
-import 'binding.dart';
 import 'binding_string.dart';
-import 'type.dart';
 import 'utils.dart';
 import 'writer.dart';