+++ lib/src/code_generator/library.dart
@@ -6,9 +6,11 @@ import 'dart:io';
 
 import 'package:cli_util/cli_util.dart';
 import 'package:ffigen/src/config_provider/config_types.dart';
+import 'package:ffigen/src/strings.dart' as strings;
 import 'package:logging/logging.dart';
 import 'package:path/path.dart' as p;
 import 'binding.dart';
+import 'imports.dart';
 import 'struc.dart';
 import 'utils.dart';
 import 'writer.dart';
@@ -31,6 +33,7 @@ class Library {
     bool dartBool = true,
     bool sort = false,
     StructPackingOverride? packingOverride,
+    Set<LibraryImport>? libraryImports,
   }) {
     /// Get all dependencies (includes itself).
     final dependencies = <Binding>{};
@@ -67,6 +70,13 @@ class Library {
     final noLookUpBindings =
         this.bindings.whereType<NoLookUpBinding>().toList();
 
+    // Add predefined imports.
+    final _libraryImports = <LibraryImport>{};
+    _libraryImports.addAll(strings.predefinedLibraryImports.values);
+    if (libraryImports != null) {
+      _libraryImports.addAll(libraryImports);
+    }
+
     _writer = Writer(
       lookUpBindings: lookUpBindings,
       noLookUpBindings: noLookUpBindings,
@@ -74,6 +84,7 @@ class Library {
       classDocComment: description,
       header: header,
       dartBool: dartBool,
+      libraryImports: _libraryImports,
     );
   }