+++ test/test_utils.dart
@@ -70,3 +70,10 @@ void logWarnings([Level level = Level.WARNING]) {
     print('${record.level.name.padRight(8)}: ${record.message}');
   });
 }
+
+void logWarningsToArray(List<String> logArr, [Level level = Level.WARNING]) {
+  Logger.root.level = level;
+  Logger.root.onRecord.listen((record) {
+    logArr.add('${record.level.name.padRight(8)}: ${record.message}');
+  });
+}