+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 5.0.0-dev.0
+- Added support for generating ABI Specific integers.
+- Breaking: removed config keys - `size-map` and `typedef-map`.
+- Added config keys - `library-imports` and `type-map`.
+
 # 4.1.2
 - Added fix for empty include list to exclude all